+++ waypoints/src/test/kotlin/de/md5lukas/waypoints/data/FolderTest.kt
@@ -4,9 +4,10 @@ import de.md5lukas.waypoints.api.Icon
 import de.md5lukas.waypoints.api.Type
 import de.md5lukas.waypoints.api.event.FolderPostDeleteEvent
 import de.md5lukas.waypoints.api.event.FolderPreDeleteEvent
+import de.md5lukas.waypoints.util.Items
+import de.md5lukas.waypoints.util.getValue
 import kotlin.test.assertEquals
 import kotlinx.coroutines.runBlocking
-import org.bukkit.Material
 import org.junit.jupiter.api.assertAll
 
 class FolderTest : TestBase() {
@@ -35,14 +36,15 @@ class FolderTest : TestBase() {
 
     folder.setName("Other name")
     folder.setDescription("Some description")
-    folder.setIcon(Icon(Material.GRASS_BLOCK, null))
+    val grassBlock = Items.GRASS_BLOCK.getValue()
+    folder.setIcon(Icon(grassBlock, null, null))
 
     folder = holder.getFolders()[0]
 
     assertAll(
         { assertEquals("Other name", folder.name) },
         { assertEquals("Some description", folder.description) },
-        { assertEquals(Icon(Material.GRASS_BLOCK, null), folder.icon) })
+        { assertEquals(Icon(grassBlock, null, null), folder.icon) })
   }
 
   @TypesNoDeath
@@ -51,10 +53,13 @@ class FolderTest : TestBase() {
 
     var folder = holder.createFolder("Test")
 
-    folder.setIcon(Icon(Material.GRASS_BLOCK, 1))
+    val grassBlock = Items.GRASS_BLOCK.getValue()
+    val customModelData = """{"floats": [ 0.2, 0.1, 200 ]}"""
+    val texture = "96775476bf1ca6c730cd9dfc8675a4f497ce7aa5d401098373c8eca177159c79"
+    folder.setIcon(Icon(grassBlock, customModelData, texture))
 
     folder = holder.getFolders()[0]
 
-    assertEquals(Icon(Material.GRASS_BLOCK, 1), folder.icon)
+    assertEquals(Icon(grassBlock, customModelData, texture), folder.icon)
   }
 }