+++ waypoints/src/test/kotlin/de/md5lukas/waypoints/data/WaypointTest.kt
@@ -6,12 +6,13 @@ import de.md5lukas.waypoints.api.WaypointsPlayer
 import de.md5lukas.waypoints.api.event.WaypointPostDeleteEvent
 import de.md5lukas.waypoints.api.event.WaypointPreDeleteEvent
 import de.md5lukas.waypoints.pointers.BeaconColor
+import de.md5lukas.waypoints.util.Items
+import de.md5lukas.waypoints.util.getValue
 import java.sql.SQLException
 import java.time.OffsetDateTime
 import java.util.*
 import kotlin.test.*
 import kotlinx.coroutines.runBlocking
-import org.bukkit.Material
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.assertAll
 import org.junit.jupiter.api.assertThrows
@@ -45,7 +46,8 @@ class WaypointTest : TestBase() {
     if (type === Type.PERMISSION) {
       waypoint.setPermission("permission")
     }
-    waypoint.setIcon(Icon(Material.GRASS_BLOCK, null))
+    val grassBlock = Items.GRASS_BLOCK.getValue()
+    waypoint.setIcon(Icon(grassBlock, null, null))
     waypoint.setBeaconColor(BeaconColor.LIGHT_GRAY)
 
     waypoint = holder.getWaypoints()[0]
@@ -58,7 +60,7 @@ class WaypointTest : TestBase() {
             assertEquals("permission", waypoint.permission)
           }
         },
-        { assertEquals(Icon(Material.GRASS_BLOCK, null), waypoint.icon) },
+        { assertEquals(Icon(grassBlock, null, null), waypoint.icon) },
         { assertEquals(BeaconColor.LIGHT_GRAY, waypoint.beaconColor) })
   }
 
@@ -68,11 +70,14 @@ class WaypointTest : TestBase() {
 
     var waypoint = holder.createWaypoint("Test", server.createLocation("world", 1, 2, 3))
 
-    waypoint.setIcon(Icon(Material.GRASS_BLOCK, 1))
+    val grassBlock = Items.GRASS_BLOCK.getValue()
+    val customModelData = """{"floats": [ 0.2, 0.1, 200 ]}"""
+    val texture = "96775476bf1ca6c730cd9dfc8675a4f497ce7aa5d401098373c8eca177159c79"
+    waypoint.setIcon(Icon(grassBlock, customModelData, texture))
 
     waypoint = holder.getWaypoints()[0]
 
-    assertEquals(Icon(Material.GRASS_BLOCK, 1), waypoint.icon)
+    assertEquals(Icon(grassBlock, customModelData, texture), waypoint.icon)
   }
 
   @Nested